Updated how ose_installer is being loaded in foreman tasks.
