+++ server/app/lib/actions/fusor/deployment/open_shift/post_install_ose.rb
@@ -10,8 +10,6 @@
 # have received a copy of GPLv2 along with this software; if not, see
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
 
-require "#{Fusor::Engine.root}/lib/modules/ose-installer/launch.rb"
-
 module Actions
   module Fusor
     module Deployment
