+++ server/lib/modules/ose_installer/files/docker-storage-setup
@@ -1,0 +1,9 @@
+# Edit this file to override any configuration options specified in
+# /usr/lib/docker-storage-setup/docker-storage-setup.
+#
+# For more details refer to "man docker-storage-setup"
+
+# Workaround for https://trello.com/c/1fhpNoVQ, use loopback
+#DEVS=/dev/vdb
+#VG=docker-vg
+
