+++ server/lib/modules/ose_installer/files/main.yml
