+++ server/lib/modules/ose_installer/files/pod_examples/hello_openshift.json
