+++ server/lib/modules/ose_installer/files/pod_examples/registry-volume.json
