+++ server/lib/modules/ose_installer/playbooks/install.yml
