+++ server/lib/modules/ose_installer/playbooks/post_install.yml
