+++ server/lib/modules/ose_installer/playbooks/setup.yml
