+++ server/lib/modules/ose_installer/templates/ansible.hosts.template
