+++ server/lib/modules/ose_installer/templates/atomic-openshift-installer.answers.cfg.yml.template
