+++ server/lib/modules/ose_installer/templates/docker-storage-setup.template
